# config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
  "/registration": {
    "post": {
      "tags": ["Authentication"],
      "description": "Регистрация нового пользователя",
      "parameters": [
        {
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "properties": {
              "username": {
                "example": "any"
              },
              "password": {
                "example": "any"
              },
              "email": {
                "example": "any"
              }
            }
          }
        }
      ],
      "responses": {
        "201": {
          "description": "Created"
        },
        "400": {
          "description": "Bad Request"
        },
        "500": {
          "description": "Internal Server Error"
        }
      }
    }
  },
  "/authentication": {
    "post": {
      "tags": ["Authentication"],
      "description": "Аутентификация пользователя",
      "parameters": [
        {
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "example": "any"
              },
              "password": {
                "example": "any"
              }
            }
          }
        }
      ],
      "responses": {
        "200": {
          "description": "OK"
        },
        "400": {
          "description": "Bad Request"
        },
        "401": {
          "description": "Unauthorized"
        },
        "500": {
          "description": "Internal Server Error"
        }
      }
    }
},
  "/logout": {
    "delete": {
      "tags": ["Authentication"],
      "description": "Log out the current user by invalidating their session or token",
      "responses": {
        "200": {
          "description": "Successfully logged out"
        },
        "401": {
          "description": "Unauthorized, user not authenticated"
        },
        "500": {
          "description": "Internal Server Error"
        }
      }
    }
  },
...
  }
}